+++ commandIns/custom-commands/模板1.js
@@ -1,0 +1,4 @@
+
+module.exports = function () {
+    let { vscodeApi, nodeApi } = this
+}