+++ commandIns/init-Configs/jiali3.js
@@ -1,0 +1,5 @@
+
+module.exports = {
+path: '请输入相对项目根目录的文件路径',
+content: ` 请输入文件内容`
+}